Corrected SRLM function and added sim scripts
I found an error in the SRLM function where the rate of extinction of a patch was being added to the equation N x for the N patches in the system akin to colonization (not just for that individual patch). I checked this and have added scripts for the simulation experiments.